+++ src/App.tsx
@@ -1,12 +1,15 @@
 import React from 'react';
+import './App.css';
 import Index from './pages/Index';
+import PerformanceMonitor from './components/PerformanceMonitor';
 
-const App = () => {
+function App() {
   return (
-    <div className="min-h-screen bg-background">
+    <div className="App">
       <Index />
+      <PerformanceMonitor />
     </div>
   );
-};
+}
 
 export default App;